Address review feedback: fix created_at semantics, add unassign UI, surface affected agents, extract shared utils
- Remove created_at reset on upsert to preserve original timestamp
- Add "None (unassigned)" option to payment method dropdown
- Wire up useRemoveAgentPaymentMethod hook for unassign flow
- Show affected agents count in delete toast message
- Extract formatBrand/formatExpiry to shared paymentMethodUtils.ts
