+++ frontend/src/lib/paymentMethodUtils.ts
@@ -1,0 +1,16 @@
+export function formatBrand(brand: string): string {
+  const brands: Record<string, string> = {
+    visa: "Visa",
+    mastercard: "Mastercard",
+    amex: "Amex",
+    discover: "Discover",
+    diners: "Diners",
+    jcb: "JCB",
+    unionpay: "UnionPay",
+  };
+  return brands[brand] ?? brand;
+}
+
+export function formatExpiry(month: number, year: number): string {
+  return `${String(month).padStart(2, "0")}/${String(year).slice(-2)}`;
+}
